requirements.txt
async-timeout==4.0.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.1
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
